Add new Palm OS target to GCC and binutils
The original Palm OS prc-tools target was coff-m68k, but support
for that was removed from GCC in 2018. Since the rest of Retro68
already uses elf32-m68k and Palm OS works very similarly to Mac
OS, it makes sense to use that too, and just convert the old SDK
glue libraries to ELF later.

Without support in libretro, this commit does not do much except
enable compilation of a compiler and tools that will eventually
be able to compile to Palm OS.