Add comments to loader script entrypoint trampoline
This code exists at the intersection of multiple arcane things—
binutils ld scripts, raw byte code for an obsolete architecture,
and a hack for early hardware revisions of that architecture. It
took an embarrassingly long time for me to figure out what exactly
was going on, so I add these comments in the hope that a future
traveller will be able to understand things more quickly.