Implement m68k COFF to ELF support in objconv
objconv does the bad thing of loading data off disk directly into
memory and then addressing it with a cast to struct, so is totally
dependent on the target machine having a compatible memory layout
and CPU endianness. It only includes support for COFF/PE, and only
ever x86 or x86_64, and makes a few assumptions in that regard too.
For these reasons, it is not a great choice, especially since Palm
OS glue libraries are pretty simple, using only one relocation type
and small .text, .data, and .bss sections. However, there is just
enough esotera in handling conversions between COFF and ELF, and in
reading GNU variant ar archives, that it did not quite make sense
to try to write a smaller independent utility (though the decision
was a marginal one). The other option of patching binutils to
reintroduce deleted coff-m68k support seemed like an even worse
idea, since whilst bfd is much more portable, it is also much more
ancient C, and binutils will require ongoing updates from upstream,
so adding non-trivial patches that would also never be accepted
upstream seemed like an even worse idea.

This patch is grotesque and unlikely to ever make it upstream; it
was written to avoid breaking the existing COFF/PE handling, but it
only targets COFF-to-ELF conversion (because it would be a silly
waste of time to do the rest of the work when this is all that
Retro68 needs). It is also an absolutely awful hack that does not
actually rearchitect objconv to do the right thing; instead, it
just swaps the bytes in memory and then swaps them back later, in
a way that is brittle and hard to rationalise about.