Add PilRC to tree
This is the latest version of PilRC that seems to exist in wide
distribution online, lightly modified to replace the ancient build
files with CMake.

There may be a slightly different patched version circulating
elsewhere, but there is no upstream maintainer for any of it, so
it can just live in the tree here from now.