Fix 64-bit soundness and other issues in PilRC
The upstream PilRC source code rotted badly enough that the UI
code would not even build against currently extant versions of
GTK 2, which is itself already obsolete. The use of obsolete APIs
has been removed to enable that code to compile again, though it
barely works.

The existing source code was not 64-bit safe and used both a bad
check to pick an appropriate type to hold pointers instead of just
using `intptr_t`, and also passed these values through `va_args`
in a way that caused things to break without the compiler knowing
about it.

Running static analysis to try to find and resolve any additional
64-bit soundness issues uncovered some other trivially fixable
issues with the code, which are now fixed. It is probable that
there are still memory leaks in at least one place, but this is a
short-running utility, so in the absence of a real incentive,
memory leaks are better than potential use-after-frees resulting
from careless, rushed fixes.