Replace old busted gzip library with libz
The existence of this old thing causes symbol confusion, leading
to crashes when system libraries try to call zlib functions. In
particular, using the native filesystem picker from FLTK will try
to use GTK+, which will try to load PNG files with libpng, which
will try to call the built-in inflate, which will crash.