Fix missing skin transparency
This only works with FLTK 1.3.3 or later on Linux *when FLTK is
compiled with `-DOPTION_ABI_VERSION=10303`* (or higher). This is
*not* the default, and most Linux distributions still ship with the
10300 ABI, which does not enable window shaping. In this case, the
window will just be decorated normally. This ABI change is 10 years
old already, so it is very frustrating that the upstream project
has still not made a new release to make this the default.